Optimisers.Descent — Type

Descent(η = 1f-1)

Classic gradient descent optimiser with learning rate η. For each parameter p and its gradient dp, this runs p -= η*dp.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Optimisers.Momentum — Type

Momentum(η = 1f-2, ρ = 9f-1)

Gradient descent optimizer with learning rate η and momentum ρ.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Optimisers.Nesterov — Type

Nesterov(η = 1f-3, ρ = 9f-1)

Gradient descent optimizer with learning rate η and Nesterov momentum ρ.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Nesterov momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Optimisers.RMSProp — Type

RMSProp(η = 1f-3, ρ = 9f-1, ϵ = eps(typeof(η)))

Optimizer using the RMSProp algorithm. Often a good choice for recurrent networks. Parameters other than learning rate generally don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.ADAM — Type

ADAM(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

ADAM optimiser.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.RADAM — Type

RADAM(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

Rectified ADAM optimizer.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.AdaMax — Type

AdaMax(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

AdaMax is a variant of ADAM based on the ∞-norm.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.OADAM — Type

OADAM(η = 1f-3, β = (5f-1, 9f-1), ϵ = eps(typeof(η)))

OADAM (Optimistic ADAM) is a variant of ADAM adding an "optimistic" term suitable for adversarial training.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.ADAGrad — Type

ADAGrad(η = 1f-1, ϵ = eps(typeof(η)))

ADAGrad optimizer. It has parameter specific learning rates based on how frequently it is updated. Parameters don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.ADADelta — Type

ADADelta(ρ = 9f-1, ϵ = eps(typeof(ρ)))

ADADelta is a version of ADAGrad adapting its learning rate based on a window of past gradient updates. Parameters don't need tuning.

Parameters

Rho (ρ): Factor by which the gradient is decayed at each time step.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.AMSGrad — Type

AMSGrad(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

The AMSGrad version of the ADAM optimiser. Parameters don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.NADAM — Type

NADAM(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

NADAM is a Nesterov variant of ADAM. Parameters don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.ADAMW — Function

ADAMW(η = 1f-3, β = (9f-1, 9.99f-1), γ = 0, ϵ = eps(typeof(η)))

ADAMW is a variant of ADAM fixing (as in repairing) its weight decay regularization.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Weight decay (γ): Decay applied to weights during optimisation.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Optimisers.AdaBelief — Type

AdaBelief(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

The AdaBelief optimiser is a variant of the well-known ADAM optimiser.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ::Float32): Constant to prevent division by zero (no need to change default)

Missing docstring.

Missing docstring for Optimisers.weightDecay. Check Documenter's build log for details.

Optimisers.OptimiserChain — Type

OptimiserChain(opts...)

Compose a chain (sequence) of optimisers so that each opt in opts updates the gradient in the order specified.