Optimisation Rules

Optimisers.Descent — Type

Descent(η = 1f-1)

Classic gradient descent optimiser with learning rate η. For each parameter p and its gradient dp, this runs p -= η*dp.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Optimisers.Momentum — Type

Momentum(η = 1f-2, ρ = 9f-1)

Gradient descent optimizer with learning rate η and momentum ρ.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Optimisers.Nesterov — Type

Nesterov(η = 1f-3, ρ = 9f-1)

Gradient descent optimizer with learning rate η and Nesterov momentum ρ.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Nesterov momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Optimisers.RMSProp — Type

RMSProp(η = 1f-3, ρ = 9f-1, ϵ = eps(typeof(η)); centred = false)

Optimizer using the RMSProp algorithm. Often a good choice for recurrent networks. Parameters other than learning rate generally don't need tuning.

Centred RMSProp is a variant which normalises gradients by an estimate their variance, instead of their second moment.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Momentum (ρ): Controls the acceleration of gradient descent in the prominent direction, in effect dampening oscillations.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Keyword centred (or centered): Indicates whether to use centred variant of the algorithm.

Missing docstring.

Missing docstring for Optimisers.ADAM. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Optimisers.RADAM. Check Documenter's build log for details.

Optimisers.AdaMax — Type

AdaMax(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

AdaMax is a variant of Adam based on the ∞-norm.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Missing docstring.

Missing docstring for Optimisers.OADAM. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Optimisers.ADAGrad. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Optimisers.ADADelta. Check Documenter's build log for details.

Optimisers.AMSGrad — Type

AMSGrad(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = eps(typeof(η)))

The AMSGrad version of the Adam optimiser. Parameters don't need tuning.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ): Constant to prevent division by zero (no need to change default)

Missing docstring.

Missing docstring for Optimisers.NADAM. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Optimisers.ADAMW. Check Documenter's build log for details.

Optimisers.AdaBelief — Type

AdaBelief(η = 1f-3, β = (9f-1, 9.99f-1), ϵ = 1e-16)

The AdaBelief optimiser is a variant of the well-known Adam optimiser.

Parameters

Learning rate (η): Amount by which gradients are discounted before updating the weights.

Decay of momentums (β::Tuple): Exponential decay for the first (β1) and the second (β2) momentum estimate.

Machine epsilon (ϵ::Float32): Constant to prevent division by zero (no need to change default)

In addition to the main course, you may wish to order some of these condiments:

Optimisers.ClipGrad — Type

ClipGrad(δ = 10f0)

Restricts every gradient component to obey -δ ≤ dx[i] ≤ δ.

See also ClipNorm.

Optimisers.ClipNorm — Type

ClipNorm(ω = 10f0, p = 2; throw = true)

Scales any gradient array for which norm(dx, p) > ω to stay at this threshold (unless p==0).

Throws an error if the norm is infinite or NaN, which you can turn off with throw = false.

See also ClipGrad.

Optimisers.WeightDecay — Type

WeightDecay(γ = 5f-4)

Decay weights by $γ$, that is, add γ .* x to the gradient x̄ which will be subtracted from x.

Typically composed with other optimisers as the first transformation in an OptimiserChain. This is equivalent to adding $L_2$ regularization with coefficient $γ$ to the loss.

Parameters

Weight decay (γ): Decay applied to weights during optimisation.

Optimisers.OptimiserChain — Type

OptimiserChain(opts...)

Compose a sequence of optimisers so that each opt in opts updates the gradient, in the order specified.

With an empty sequence, OptimiserChain() is the identity, so update! will subtract the full gradient from the parameters. This is equivalent to Descent(1).

Example

julia> o = OptimiserChain(ClipGrad(1), Descent(0.1));

julia> m = (zeros(3),);

julia> s = Optimisers.setup(o, m)
(Leaf(OptimiserChain(ClipGrad{Int64}(1), Descent{Float64}(0.1)), [nothing, nothing]),)

julia> Optimisers.update(s, m, ([0.3, 1, 7],))[2]  # clips before discounting
([-0.03, -0.1, -0.1],)

Model Interface

Optimisers.setup — Function

Optimisers.setup(rule, model) -> tree

Initialises the given optimiser for every trainable parameter within the model. Returns a tree of the relevant states, which must be passed to update or update!.

Example

julia> Optimisers.setup(Descent(0.1f0), (x = rand(3), y = (true, false), z = tanh))
(x = Leaf(Descent{Float32}(0.1), nothing), y = (nothing, nothing), z = nothing)

Optimisers.update — Function

Optimisers.update(tree, model, gradient) -> (tree, model)

Uses the optimiser and the gradient to change the trainable parameters in the model. Returns the improved model, and the optimiser states needed for the next update. The initial tree of states comes from setup.

See also update!, which will be faster for models of ordinary Arrays or CuArrays.

Example

julia> m = (x = Float32[1,2,3], y = tanh);

julia> t = Optimisers.setup(Descent(0.1f0), m)
(x = Leaf(Descent{Float32}(0.1), nothing), y = nothing)

julia> g = (x = [1,1,1], y = nothing);  # fake gradient

julia> Optimisers.update(t, m, g)
((x = Leaf(Descent{Float32}(0.1), nothing), y = nothing), (x = Float32[0.9, 1.9, 2.9], y = tanh))

Optimisers.update! — Function

Optimisers.update!(tree, model, gradient) -> (tree, model)

Uses the optimiser and the gradient to change the trainable parameters in the model. Returns the improved model, and the optimiser states needed for the next update. The initial tree of states comes from setup.

This is used in exactly the same manner as update, but because it may mutate arrays within the old model (and the old state), it will be faster for models of ordinary Arrays or CuArrays. However, you should not rely on the old model being fully updated.

Calling Functors.@functor on your model's layer types by default causes the optimiser to act on all suitable fields. To restrict this, define trainable:

Optimisers.trainable — Function

trainable(x::Layer) -> NamedTuple

This should be overloaded to make optimisers ignore some fields of every Layer, which would otherwise contain trainable parameters. (Elements such as functions and sizes are always ignored.)

The default is Functors.children(x), usually a NamedTuple of all fields, and trainable(x) must contain a subset of these.

Such restrictions are also obeyed by this function for flattening a model:

Optimisers.destructure — Function

destructure(model) -> vector, reconstructor

Copies all trainable, isnumeric parameters in the model to a vector, and returns also a function which reverses this transformation. Differentiable.

Example

julia> v, re = destructure((x=[1.0, 2.0], y=(sin, [3 + 4im])))
(ComplexF64[1.0 + 0.0im, 2.0 + 0.0im, 3.0 + 4.0im], Restructure(NamedTuple, ..., 3))

julia> re([3, 5-im, 7+11im])
(x = [3.0, 5.0], y = (sin, ComplexF64[7.0 + 11.0im]))

Optimisers.Restructure — Type

Restructure(Model, ..., length)

This is what destructure returns, and re(p) will re-build the model with new parameters from vector p. If the model is callable, then re(x, p) == re(p)(x).

Example

julia> using Flux, Optimisers

julia> _, re = destructure(Dense([1 2; 3 4], [0, 0], sigmoid))
([1, 3, 2, 4, 0, 0], Restructure(Dense, ..., 6))

julia> m = re(-4:1)
Dense(2, 2, σ)      # 6 parameters

julia> m([0.2, 0.3]) ≈ re([0.2, 0.3], -4:1)
true

Rule Definition

Optimisers.apply! — Function

Optimisers.apply!(rule::RuleType, state, parameters, gradient) -> (state, gradient)

This defines the action of any optimisation rule. It should return the modified gradient which will be subtracted from the parameters, and the updated state (if any) for use at the next iteration, as a tuple (state, gradient).

For efficiency it is free to mutate the old state, but only what is returned will be used. Ideally this should check iswriteable(x), which the built-in rules do via @...

The initial state is init(rule::RuleType, parameters).

Example

julia> Optimisers.init(Descent(0.1), Float32[1,2,3]) === nothing
true

julia> Optimisers.apply!(Descent(0.1), nothing, Float32[1,2,3], [4,5,6])
(nothing, Base.Broadcast.Broadcasted{Base.Broadcast.DefaultArrayStyle{1}}(*, ([4, 5, 6], 0.1f0)))

Optimisers.init — Function

Optimisers.init(rule::RuleType, parameters) -> state

Sets up the initial state for a given optimisation rule, and an array of parameters. This and apply! are the two functions which any new optimisation rule must define.

Examples

julia> Optimisers.init(Descent(), Float32[1,2,3])  # is `nothing`

julia> Optimisers.init(Momentum(), [1.0, 2.0])
2-element Vector{Float64}:
 0.0
 0.0

Optimisers.@.. — Macro

@.. x = x + y

Sometimes in-place broadcasting macro, for use in apply! rules. If iswriteable(x) then it is just @. x = rhs, but if not, it becomes x = @. rhs.

Optimisers.@lazy — Macro

x = @lazy y + z

Lazy broadcasting macro, for use in apply! rules. It broadcasts like @. but does not materialise, returning a Broadcasted object for later use. Beware that mutation of arguments will affect the result, and that if it is used in two places, work will be done twice.